+++ src/features/defi/providers/yearn/components/YearnManager/Withdraw/WithdrawCommon.ts
@@ -1,0 +1,6 @@
+export enum WithdrawPath {
+  Withdraw = '/',
+  Confirm = '/confirm',
+  ConfirmSettings = '/confirm/settings',
+  Status = '/status',
+}
