+++ src/features/defi/providers/yearn/components/YearnManager/Withdraw/WithdrawContext.ts
@@ -1,0 +1,10 @@
+import { createContext } from 'react'
+
+import { YearnWithdrawActions, YearnWithdrawState } from './WithdrawReducer'
+
+export interface IWithdrawContext {
+  state: YearnWithdrawState | null
+  dispatch: React.Dispatch<YearnWithdrawActions> | null
+}
+
+export const WithdrawContext = createContext<IWithdrawContext>({ state: null, dispatch: null })
